hub/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.0,<3.0
pyserial>=3.5,<4.0
httpx>=0.27,<1.0
